Behavior of the generator in case of overwriting existing files is now manageable

+++ codegen-generator/src/main/java/de/chrlembeck/codegen/generator/output/FileOutput.java
@@ -5,6 +5,7 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.OutputStreamWriter;
 import java.io.Writer;
+import java.nio.file.FileAlreadyExistsException;
 import java.nio.file.Path;
 import java.util.Map;
 import java.util.TreeMap;
@@ -58,12 +59,46 @@ public class FileOutput implements GeneratorOutput {
             if (!directory.toFile().exists()) {
                 directory.toFile().mkdirs();
             }
-            final FileOutputStream fileOut = new FileOutputStream(path.toFile());
-            final BufferedOutputStream bufOut = new BufferedOutputStream(fileOut);
-            writer = new OutputStreamWriter(bufOut, prefs.getCharsetForChannel(channelName));
+            if (path.toFile().exists() && keepExisting(channelName, prefs, path)) {
+                writer = new NullWriter();
+            } else {
+                final FileOutputStream fileOut = new FileOutputStream(path.toFile());
+                final BufferedOutputStream bufOut = new BufferedOutputStream(fileOut);
+                writer = new OutputStreamWriter(bufOut, prefs.getCharsetForChannel(channelName));
+            }
             writers.put(channelName, writer);
         }
         return writer;
+    }
+
+    /**
+     * Prüft, ob die bestehende Datei ersetzt oder behalten werden soll.
+     * 
+     * @param channelName
+     *            Name des Kanals, in den geschrieben werden soll.
+     * @param prefs
+     *            Einstellungen für die Ausgabe des Generators.
+     * @param path
+     *            Pfad auf die zu ersetzende Datei
+     * @return {@code true} falls die Datei bestehen bleiben soll, {@code false} falls die Datei ersetzt werden soll.
+     * @throws FileAlreadyExistsException
+     *             Falls die Einstellungen beim Antreffen einer bestehenden Datei eine Exception erwarten.
+     */
+    private boolean keepExisting(final String channelName, final OutputPreferences prefs, final Path path)
+            throws FileAlreadyExistsException {
+        final OverwritePreferences overwritePreferences = prefs.getOverwritePreferencesForChannel(channelName);
+        if (overwritePreferences == OverwritePreferences.THROW_EXCEPTION) {
+            final String message = "Die zu schreibende Datei existiert bereits und soll nicht überschrieben werden: "
+                    + path;
+            LOGGER.error(message);
+            throw new FileAlreadyExistsException(
+                    message);
+        } else if (overwritePreferences == OverwritePreferences.KEEP_EXISTING) {
+            LOGGER.warn("Die Ausgabedatei existiert bereits und wird aufgrund der Einstellung "
+                    + overwritePreferences + " nicht verändert: " + path);
+            return true;
+        }
+        return false;
     }
 
     /**
@@ -88,6 +123,38 @@ public class FileOutput implements GeneratorOutput {
             } catch (final IOException e) {
                 LOGGER.error("Writer konnte nicht geschlossen werden.", e);
             }
+        }
+    }
+
+    /**
+     * Hilfsklasse für das Verhalten des Generators, wenn er eine existierende Datei nicht überschreiben soll.
+     * 
+     * @author Christoph Lembeck
+     */
+    static class NullWriter extends Writer {
+
+        /**
+         * Funktionslos.
+         */
+        @Override
+        public void write(final char[] cbuf, final int off, final int len) throws IOException {
+            // do nothing
+        }
+
+        /**
+         * Funktionslos.
+         */
+        @Override
+        public void flush() throws IOException {
+            // do nothing
+        }
+
+        /**
+         * Funktionslos.
+         */
+        @Override
+        public void close() throws IOException {
+            // do nothing
         }
     }
 }
